Cambios, parametros y estructuración
