+++ app/services/notification/I_notification_type.py
@@ -4,5 +4,5 @@ from abc import ABC, abstractmethod
 # Clase abstracta para el servicio de notificaciones por correo electronico
 class NotificationService(ABC):
   @abstractmethod
-  def send(self, to: str, subject: str, message: str) -> bool:
+  def send(self, data: dict) -> bool:
     pass
