--- app/services/notification/service.py
@@ -1,2 +1,0 @@
-class service():
-
